requirements.txt
Flask
flask-bootstrap
flask-wtf
pymongo
PyYaml
